Add "sharp" biased WH 31-EDO
The 31-EDO Wicki-Hayden split layouts are a little inconvenient, because even
with the right hand fitting, it gets split unnaturally.  The left hand is even
worse, because there are just notes missing.

This is the same as the main WH split layout, but with everything shifted a bit
over to the left.  This helps when playing music that is "sharp biased", as it
keeps the double sharps on both sides together, at the expense of splitting the
double flats.  With this layout and the regular WH split, most classical music
can just be played in a given layout.  More microtonal-focused music may require
additional shifts, possibly dynamically changed, to be comfortable to play.